implement calculateQuantityByInvestment and calculateActualValue methods; update AssetRepository to return BigDecimal; add new endpoints in AssetController

+++ backend/project/src/main/java/com/backend/project/domain/repository/AssetRepository.java
@@ -16,6 +16,6 @@ public interface AssetRepository {
     void deleteBySymbolAndUserId(String symbol, UUID userId);
     List<AssetEntity> findAllByUserId(UUID userId);
 
-    Long getQuantityByUser(UUID userId);
+    BigDecimal getQuantityByUser(UUID userId);
     BigDecimal getActualValue(UUID userId);
 }
